Add files via upload
This script calls and runs the Adaptive Weights Smoothing with Optimized Metrics (AWSOM) algorithm (awsom.R). 
The results will be the p-value activity map, z-values and the filtered beta-value activity map.

It requires the path to the following files obtained by the GLM analysis:
- copePath: beta map, also called contrast parameter (COPE) in FSL
- residualsPath: residuals
 - varianceCopePath: COPE variance
- maskPath: binary brain mask
- numRegressors: indicate with an integer the number of regressors used in the analysis
- outputPath: path to the output folder


To run the awsom.R script you need to have R installed,  as well as the following libraries:
- fmri: https://cran.r-project.org/web/packages/fmri/index.html
- RNifti: https://cran.r-project.org/web/packages/RNifti/index.html
- oro.nifti : https://cran.r-project.org/web/packages/oro.nifti/index.html
In addition you need to install the fmri_1.9.12.3.tar.gz file from:
https://github.com/IgorTellez/AWSOM/blob/e97913dd646b3d0ae9b66983f7e9228ad56d5736/fmri_1.9.12.3.tar.gz


The FSL toolbox needs to be installed to convert the map from p-values to z-values.